+++ Java/game/src/main/java/com/yan/game/service/IResultService.java
@@ -5,7 +5,7 @@ import com.yan.game.dto.InfoDto;
 import java.util.List;
 
 public interface IResultService {
-    List<InfoDto> getUserInfo(Integer id);
-    List<InfoDto> getLevelInfo(Integer id);
+    List<InfoDto> getUserInfo(Long id);
+    List<InfoDto> getLevelInfo(Long id);
     void setInfo(InfoDto infoDto);
 }
